Move workers to their own interface project
