+++ Refresh.GameServer/RefreshGameServer.cs
@@ -23,13 +23,14 @@ using Refresh.GameServer.Authentication;
 using Refresh.Database.Models.Authentication;
 using Refresh.Database;
 using Refresh.GameServer.Middlewares;
-using Refresh.GameServer.Services;
 using Refresh.Database.Models.Users;
-using Refresh.GameServer.Workers;
 using Refresh.Database.Models.Levels;
 using Refresh.Interfaces.APIv3;
 using Refresh.Interfaces.APIv3.Documentation;
 using Refresh.Interfaces.Game;
+using Refresh.Interfaces.Workers;
+using Refresh.Interfaces.Workers.RequestTracking;
+using Refresh.Interfaces.Workers.Workers;
 
 namespace Refresh.GameServer;
