+++ Refresh.Interfaces.Workers/Refresh.Interfaces.Workers.csproj
@@ -8,7 +8,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\Refresh.Core\Refresh.Core.csproj" />
+        <ProjectReference Include="..\Refresh.Core\Refresh.Core.csproj" />
+        <ProjectReference Include="..\Refresh.Interfaces.APIv3\Refresh.Interfaces.APIv3.csproj" />
     </ItemGroup>
 
 </Project>
