+++ Refresh.sln
@@ -27,6 +27,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Refresh.Interfaces.APIv3", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Refresh.Interfaces.Game", "Refresh.Interfaces.Game\Refresh.Interfaces.Game.csproj", "{9EACC405-0340-46C9-8437-68AC71AB05CF}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Refresh.Interfaces.Workers", "Refresh.Interfaces.Workers\Refresh.Interfaces.Workers.csproj", "{0FD65A5F-F20F-4A30-B486-B21146686FA6}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -110,6 +112,14 @@ Global
 		{9EACC405-0340-46C9-8437-68AC71AB05CF}.DebugLocalBunkum|Any CPU.Build.0 = Debug|Any CPU
 		{9EACC405-0340-46C9-8437-68AC71AB05CF}.DebugPostgres|Any CPU.ActiveCfg = Debug|Any CPU
 		{9EACC405-0340-46C9-8437-68AC71AB05CF}.DebugPostgres|Any CPU.Build.0 = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.DebugLocalBunkum|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.DebugLocalBunkum|Any CPU.Build.0 = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.DebugPostgres|Any CPU.ActiveCfg = Debug|Any CPU
+		{0FD65A5F-F20F-4A30-B486-B21146686FA6}.DebugPostgres|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{A3D76B31-8732-4134-907B-F36773DF70D3} = {AA547E95-555C-470E-995A-D66E56660542}
